command: update edit-validation-sets to craft-cli
This is a minimal code change migration from Snapcraft's
snapcraft_legacy module to snapcraft. Adaptations to use craft-cli for
output has been made, as well as updates to how file edit takes place by
 migrating it to use Path.

The tests have been updated to follow the newer conventions of the code
base, but the gist of them has not changed at all.

Lastly, the only UI change is that validation sets have been separated
into their own command group.
